Refactor TimespanInput and add Preview component for enhanced EDTF handling
- Updated TimespanInput component to improve EDTF value handling and patch application logic.
- Introduced a new Preview component for better visualization of Timespan values.
- Enhanced error handling for invalid EDTF values and improved conditional rendering.
- Refactored import paths and updated dependencies in package.json for consistency.
