+++ packages/sanity-plugin-timespan-input/package.json
@@ -50,17 +50,16 @@
   },
   "dependencies": {
     "@date-fns/tz": "1.2.0",
-    "@date-fns/utc": "^2.1.0",
     "@sanity/icons": "^3.5.6",
     "@sanity/incompatible-plugin": "^1.0.5",
     "@sanity/ui": "^1.8.2",
-    "@types/styled-components": "^5.1.28",
     "date-fns": "4.1.0",
     "edtf": "^4.6.0"
   },
   "devDependencies": {
     "@sanity/pkg-utils": "^6.12.2",
     "@sanity/plugin-kit": "^4.0.19",
+    "@types/styled-components": "^5.1.28",
     "@types/react": "^18.2.21",
     "@typescript-eslint/eslint-plugin": "^7.0.0",
     "@typescript-eslint/parser": "^7.0.0",
@@ -91,16 +90,6 @@
   },
   "eslintConfig": {
     "parser": "sanipack/babel/eslint-parser",
-    "rules": {
-      "import/no-unresolved": [
-        2,
-        {
-          "ignore": [
-            "^(all|part):"
-          ]
-        }
-      ]
-    },
     "extends": [
       "sanity",
       "sanity/react",
